feat: Add option to negate bits written to ByteOutputStream (#12050)
Summary:

In PrestoSerializer we write the null bits to a ByteOutputStream. Presto's serialization format inverts the null
flags from what Velox uses. This change adds an option to ByteOutputStream to do that negation on flush,
rather than requiring the user to negate them when calling append.  This makes it easier to write ranges of null
bits without an additional copy, and reduces the level of mental gymnastics needed to reason about the null 
flags in the serializer. (It's analogous to the option ByteOutputStream already has to reverse the order
of bits within a byte, again because of Presto's serialization format for null flags).

Reviewed By: xiaoxmeng

Differential Revision: D67994045